[Bronze II] Title: 숫자, Time: 16 ms, Memory: 2020 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/10093. 숫자/숫자.cc
@@ -6,13 +6,13 @@ int main(int argc, char** argv) {
     ios::sync_with_stdio(0);
     cin.tie(0);
     
-    int a, b;
+    long long a, b;
     cin >> a >> b;
     if (a > b) swap(a, b);
     if (a == b || a == b + 1) cout << 0;
     else {
         cout << b - a - 1 << "\n";
-        for (int i = a + 1; i < b; i++) {
+        for (long long i = a + 1; i < b; i++) {
             cout << i << ' ';
         }
     }
